Split datetime and duration rules vs. number and amount rules in EN grammar.
