+++ grammar/en/src/rules_amount.rs
@@ -1,0 +1,211 @@
+use rustling::*;
+use rustling_ontology_values::dimension::*;
+use rustling_ontology_values::dimension::Precision::*;
+use rustling_ontology_values::helpers;
+
+pub fn rules_percentage(b: &mut RuleSetBuilder<Dimension>) -> RustlingResult<()> {
+    b.rule_2("<number> per cent",
+        number_check!(),
+        b.reg(r"(?:%|p\.c\.|per ?cents?)")?,
+        |number, _| Ok(PercentageValue(number.value().value()))
+    );
+    Ok(())
+}
+
+pub fn rules_finance(b: &mut RuleSetBuilder<Dimension>) -> RustlingResult<()> {
+    b.rule_2("intersect (X cents)",
+             amount_of_money_check!(),
+             amount_of_money_check!(|money: &AmountOfMoneyValue| money.unit == Some("cent")),
+             |a, b| helpers::compose_money(a.value(), b.value()));
+    b.rule_3("intersect (and X cents)",
+             amount_of_money_check!(),
+             b.reg(r#"and"#)?,
+             amount_of_money_check!(|money: &AmountOfMoneyValue| money.unit == Some("cent")),
+             |a, _, b| helpers::compose_money(&a.value(), &b.value()));
+    b.rule_2("intersect",
+             amount_of_money_check!(),
+             number_check!(),
+             |a, b| helpers::compose_money_number(&a.value(), &b.value()));
+    b.rule_3("intersect (and number)",
+             amount_of_money_check!(),
+             b.reg(r#"and"#)?,
+             number_check!(),
+             |a, _, b| helpers::compose_money_number(&a.value(), &b.value()));
+    b.rule_1_terminal("$",
+                      b.reg(r#"\$|dollars?"#)?,
+                      |_| Ok(MoneyUnitValue { unit: Some("$") })
+    );
+    b.rule_1_terminal("USD",
+                      b.reg(r#"us[d\$]|(?:us|american) dollars?|bucks?"#)?,
+                      |_| Ok(MoneyUnitValue { unit: Some("USD") })
+    );
+    b.rule_1_terminal("AUD",
+                      b.reg(r#"au[d\$]|australian dollars?"#)?,
+                      |_| Ok(MoneyUnitValue { unit: Some("AUD") })
+    );
+    b.rule_1_terminal("CAD",
+                      b.reg(r#"cad|canadian dollars?"#)?,
+                      |_| Ok(MoneyUnitValue { unit: Some("CAD") })
+    );
+    b.rule_1_terminal("HKD",
+                      b.reg(r#"hkd|hk dollars?|hong[- ]?kong dollars?"#)?,
+                      |_| Ok(MoneyUnitValue { unit: Some("HKD") })
+    );
+    b.rule_1_terminal("EUR",
+                      b.reg(r#"€|(?:[e€]uro?s?)"#)?,
+                      |_| Ok(MoneyUnitValue { unit: Some("EUR") })
+    );
+    b.rule_1_terminal("£",
+                      b.reg(r#"£|pounds?"#)?,
+                      |_| Ok(MoneyUnitValue { unit: Some("£") })
+    );
+    b.rule_1_terminal("GBP",
+                      b.reg(r#"gbp|(?:sterling|british) pounds?|sterlings?|quids?"#)?,
+                      |_| Ok(MoneyUnitValue { unit: Some("GBP") })
+    );
+    b.rule_1_terminal("CHF",
+                      b.reg(r#"chf|swiss francs?"#)?,
+                      |_| Ok(MoneyUnitValue { unit: Some("CHF") })
+    );
+    b.rule_1_terminal("KR",
+                      b.reg(r#"kroner?|crowns?|kr"#)?,
+                      |_| Ok(MoneyUnitValue { unit: Some("KR") })
+    );
+    b.rule_1_terminal("DKK",
+                      b.reg(r#"dkk|danish (?:kroner?|crowns?)"#)?,
+                      |_| Ok(MoneyUnitValue { unit: Some("DKK") })
+    );
+    b.rule_1_terminal("NOK",
+                      b.reg(r#"nok|norwegian (?:kroner?|crowns?)"#)?,
+                      |_| Ok(MoneyUnitValue { unit: Some("NOK") })
+    );
+    b.rule_1_terminal("SEK",
+                      b.reg(r#"sek|swedish (?:krona|kronor|crowns?)"#)?,
+                      |_| Ok(MoneyUnitValue { unit: Some("SEK") })
+    );
+    b.rule_1_terminal("RUB",
+                      b.reg(r#"(?:russian )?rubles?|rub"#)?,
+                      |_| Ok(MoneyUnitValue { unit: Some("RUB") })
+    );
+    b.rule_1_terminal("INR",
+                      b.reg(r#"inr|rs\.?|(?:indian )?rupees?"#)?,
+                      |_| Ok(MoneyUnitValue { unit: Some("INR") })
+    );
+    b.rule_1_terminal("JPY",
+                      b.reg(r#"jpy|yens?"#)?,
+                      |_| Ok(MoneyUnitValue { unit: Some("JPY") })
+    );
+    b.rule_1_terminal("CNY",
+                      b.reg(r#"cny|cnh|rmb|yuans?|renmimbis?"#)?,
+                      |_| Ok(MoneyUnitValue { unit: Some("CNY") })
+    );
+    b.rule_1_terminal("¥",
+                      b.reg(r#"¥"#)?,
+                      |_| Ok(MoneyUnitValue { unit: Some("¥") })
+    );
+    b.rule_1_terminal("KRW",
+                      b.reg(r#"₩|krw|(?:south[- ]?)?korean wons?|wons?"#)?,
+                      |_| Ok(MoneyUnitValue { unit: Some("KRW") })
+    );
+    b.rule_1_terminal("฿",
+                      b.reg(r#"฿|bitcoins?"#)?,
+                      |_| Ok(MoneyUnitValue { unit: Some("฿") })
+    );
+    b.rule_1_terminal("cent",
+                      b.reg(r#"cents?|penn(?:y|ies)|c|¢"#)?,
+                      |_| Ok(MoneyUnitValue { unit: Some("cent") })
+    );
+    b.rule_2("<unit> <amount>",
+             money_unit!(),
+             number_check!(),
+             |a, b| {
+                 Ok(AmountOfMoneyValue {
+                     value: b.value().value(),
+                     unit: a.value().unit,
+                     ..AmountOfMoneyValue::default()
+                 })
+             });
+    b.rule_2("<amount> <unit>",
+             number_check!(),
+             money_unit!(),
+             |a, b| {
+                 Ok(AmountOfMoneyValue {
+                     value: a.value().value(),
+                     unit: b.value().unit,
+                     ..AmountOfMoneyValue::default()
+                 })
+             });
+    b.rule_2("about <amount-of-money>",
+             b.reg(r#"(?:about|approx(?:\.|imately)?|close to|near(?: to)?|around|almost)"#)?,
+             amount_of_money_check!(),
+             |_, a| {
+                 Ok(AmountOfMoneyValue {
+                     precision: Approximate,
+                     ..a.value().clone()
+                 })
+             });
+    b.rule_2("exactly <amount-of-money>",
+             b.reg(r#"exactly|precisely"#)?,
+             amount_of_money_check!(),
+             |_, a| {
+                 Ok(AmountOfMoneyValue {
+                     precision: Exact,
+                     ..a.value().clone()
+                 })
+             });
+    Ok(())
+}
+
+pub fn rules_temperature(b: &mut RuleSetBuilder<Dimension>) -> RustlingResult<()> {
+    b.rule_1("number as temp",
+             number_check!(),
+             |a| {
+                 Ok(TemperatureValue {
+                     value: a.value().value(),
+                     unit: None,
+                     latent: true,
+                 })
+             });
+    b.rule_2("<latent temp> degrees",
+             temperature_check!(|temp: &TemperatureValue| temp.latent),
+             b.reg(r#"(?:deg(?:ree?)?s?\.?)|°"#)?,
+             |a, _| {
+                 Ok(TemperatureValue {
+                     value: a.value().value,
+                     unit: Some("degree"),
+                     latent: true,
+                 })
+             });
+    b.rule_2("<temp> Celcius",
+             temperature_check!(|temp: &TemperatureValue| temp.latent),
+             b.reg(r#"centigrade|c(?:el[cs]?(?:ius)?)?\.?"#)?,
+             |a, _| {
+                 Ok(TemperatureValue {
+                     value: a.value().value,
+                     unit: Some("celsius"),
+                     latent: false,
+                 })
+             });
+    b.rule_2("<temp> Fahrenheit",
+             temperature_check!(|temp: &TemperatureValue| temp.latent),
+             b.reg(r#"f(?:ah?rh?eh?n(?:h?eit)?)?\.?"#)?,
+             |a, _| {
+                 Ok(TemperatureValue {
+                     value: a.value().value,
+                     unit: Some("fahrenheit"),
+                     latent: false,
+                 })
+             });
+    b.rule_2("<temp> Kelvin",
+             temperature_check!(),
+             b.reg(r#"k(?:elvin)?\.?"#)?,
+             |a, _| {
+                 Ok(TemperatureValue {
+                     value: a.value().value,
+                     unit: Some("kelvin"),
+                     latent: false,
+                 })
+             });
+    Ok(())
+}
+
